+++ reportsMicroservice/src/main/java/com/reportsMicroservice/demo/repository/Time_offRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface Time_offRepository extends JpaRepository<Time_off, Long> {
+public interface Time_offRepository extends JpaRepository<Time_off, Integer> {
 }
